refactor: Remove fixtures to conftest
