--- tests/fixtures/accounts.py
@@ -1,19 +1,0 @@
-import pytest
-
-
-# Hi Alice
-@pytest.fixture(scope="session")
-def alice(accounts):
-    yield accounts[0]
-
-
-# Hey-a Bob
-@pytest.fixture(scope="session")
-def bob(accounts):
-    yield accounts[1]
-
-
-# Don't forget Charlie
-@pytest.fixture(scope="session")
-def charlie(accounts):
-    yield accounts[2]
